fix: Don't delete explorer ship

+++ src/job/job.service.ts
@@ -45,8 +45,14 @@ export class JobService extends MongooseRepository<Job> {
       if (system.upgrade === 'unexplored' || system.upgrade === 'explored') {
         const fleets = await this.fleetService.findAll({empire: empire._id, location: system._id});
         const ships = await this.shipService.findAll({fleet: {$in: fleets.map(f => f._id)}});
-        const ship = this.jobLogicService.checkFleet(system.upgrade === 'unexplored' ? 'explorer' : 'colonizer', fleets, ships);
-        await this.shipService.deleteOne(ship._id);
+        if (system.upgrade === 'unexplored') {
+          // check for an explorer
+          this.jobLogicService.checkFleet('explorer', fleets, ships);
+        } else if (system.upgrade === 'explored') {
+          // check and delete colonizer
+          const ship = this.jobLogicService.checkFleet('colonizer', fleets, ships);
+          await this.shipService.deleteOne(ship._id);
+        }
       }
     }
     let time: number | undefined;
